add completed update resiudals API (gcps residuals and total RMSE) with excpetion handling and unit tests; convert source coords from int to double;

+++ src/main/java/com/amine/pfe/georef_module/entity/Gcp.java
@@ -21,10 +21,10 @@ public class Gcp {
     private GeorefImage image;
 
     @Column(nullable = false)
-    private int sourceX;
+    private Double sourceX;
 
     @Column(nullable = false)
-    private int sourceY;
+    private Double sourceY;
 
     @Column(nullable = false)
     private Double mapX;
